feat: add `StateSync` component
