+++ crates/rust-client/src/components/mod.rs
@@ -1,0 +1,1 @@
+pub mod sync_state;
