+++ crates/rust-client/src/lib.rs
@@ -9,6 +9,7 @@ pub use alloc::boxed::Box;
 extern crate std;
 
 pub mod accounts;
+pub mod components;
 pub mod notes;
 pub mod rpc;
 pub mod store;
